+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 * Reusable dataset for testing hashed versions
+* Exceptions\E4xx_UnsupportedOperation - added
 * HashedVersions\Parsers\ParseHashedVersion - added
 * HashedVersions\Values\HashedVersion - added